reorganize shared styles and components bw Select and AsyncSelect
