+++ components/src/Select/Select.js
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import WindowedSelect, { components as selectComponents } from "react-windowed-select";
+import WindowedSelect from "react-windowed-select";
 import { useTranslation } from "react-i18next";
 import { ThemeContext } from "styled-components";
 import { Field } from "../Form";
@@ -8,58 +8,9 @@ import { InlineValidation } from "../Validation";
 import customStyles from "./customReactSelectStyles";
 import { SelectPropTypes, SelectDefaultProps } from "./Select.type";
 import { SelectOption } from "./SelectOption";
+import { Control, MultiValue, ClearIndicator, SelectContainer, Menu, Input } from "./SelectComponents";
 
-const Control = props => {
-  // eslint-disable-next-line react/prop-types
-  const { isFocused } = props;
-  return (
-    <div data-testid="select-control">
-      <selectComponents.Control className={isFocused ? "nds-select--is-focused" : null} {...props} />
-    </div>
-  );
-};
-
-const MultiValue = props => {
-  return (
-    <div data-testid="select-multivalue">
-      <selectComponents.MultiValue {...props} />
-    </div>
-  );
-};
-
-const ClearIndicator = props => {
-  return (
-    <div data-testid="select-clear">
-      <selectComponents.ClearIndicator {...props} />
-    </div>
-  );
-};
-
-const Menu = props => {
-  return (
-    <div data-testid="select-dropdown">
-      <selectComponents.Menu {...props} />
-    </div>
-  );
-};
-
-const SelectContainer = props => {
-  return (
-    <div data-testid="select-container">
-      <selectComponents.SelectContainer {...props} />
-    </div>
-  );
-};
-
-const Input = props => {
-  return (
-    <div data-testid="select-input">
-      <selectComponents.Input {...props} />
-    </div>
-  );
-};
-
-const getOption = (options, value) => {
+export const getOption = (options, value) => {
   if (value == null || value === "") return value;
 
   return options.find(o => o.value === value);
@@ -156,7 +107,7 @@ const ReactSelect = ({
             ClearIndicator,
             SelectContainer,
             Menu,
-            Input,
+            Input: Input({ required }),
             ...components
           }}
           aria-label={ariaLabel}
